feat(ui): rebuild REPL chrome, slash popup, vim, keymap, compact parity (#115)
- REPL chrome: 3-row header (mascot + provider/model/cwd) replaces the
  old header_bar; chat layout pre-computes content height so short
  transcripts pin the prompt under the last message and long ones
  auto-scroll the prompt to the latest, matching the TS REPL frame.
  User rows tint full-width via Line bg; assistant text gets a coloured
  turn marker shared by streaming and finalised renders.
- Slash command popup: CommandRegistry snapshot now reaches TUI state at
  startup and re-emits on /reload-plugins via a new
  AvailableCommandsRefreshed event — the popup was previously silent in
  production. Ranker (exact > prefix > contains > tight-subsequence with
  span cap) replaces the contains-only filter; sorted snapshot stops
  HashMap randomness shuffling the popup; popup width is content-driven
  and UnicodeWidthStr-aware so CJK / wide-emoji descriptions don't clip.
- Help overlay split into help.rs (shortcuts) + help_slash.rs (commands),
  so the command surface in help and the / popup share one source.
- Vim mode wired into a working modal editor: motions, operators, text
  objects, transitions and a new wiring layer; the prior input internals
  are replaced by a dedicated textarea widget with its own tests.
- Keymap system: typed entries + module + companion tests replace the
  ad-hoc bindings; keybinding_bridge routes through the entry table.
- Compaction parity with TS: full + partial flows use structured summary
  attempts (raw kept for hooks), partial / session-memory restoration is
  driven at the query layer, and prompt-too-long retry + partial-from
  ordering now match TS semantics.
- Tool result budget: tool_result_storage gains size tracking and
  eviction; MCP tool result rendering and shell_render adjusted to the
  new shape, with new orchestration and session_start_hooks tests.
- MCP wiring: services/mcp/client and the cli mcp_handle_adapter
  rewritten with companion tests; protocol + tui_only notification
  handlers and the SDK protocol schemas updated for the new shape.

+++ coco-rs/app/tui/src/autocomplete/symbol_search.rs
@@ -18,7 +18,8 @@ pub enum SymbolSearchEvent {
     /// Search results ready.
     SearchResult {
         query: String,
-        start_pos: i32,
+        /// Byte offset where the `@#` trigger started (sentinel).
+        start_pos: usize,
         suggestions: Vec<SuggestionItem>,
     },
 }
@@ -39,7 +40,7 @@ impl SymbolSearchManager {
     }
 
     /// Schedule a debounced search.
-    pub fn search(&mut self, query: String, start_pos: i32) {
+    pub fn search(&mut self, query: String, start_pos: usize) {
         if let Some(handle) = self.pending.take() {
             handle.abort();
         }
